Data for transition model

Create data used to simulate health state transitions with a continuous time state transition model (CTSTM) given the parameters from the multi-state NMA (i.e., params_mstate_nma). The included variables are a function of the selected treatment sequences and the modeled patients.

create_transmod_data(struct, trans_mat, patients)

Arguments

struct
A model_structure object.

trans_mat
A transition matrix as returned by create_trans_mat.

patients
A data table returned from create_patients.

Value

An object of class "expanded_hesim_data" from the hesim package, which is a data table with one observation for each treatment strategy (i.e., treatment sequence), patient, and transition combination. The model structure (struct) is stored as a "model_structure" attribute and the transition matrix (trans_mat) is stored as a "trans_mat" attribute.

See also