Parameters for transition model

Extract parameters from a multi-state NMA for use with the data table returned by create_transmod_data, which are used to simulate health state transitions with a continuous time state transition model (CTSTM).

create_transmod_params(params, data, check_covs = FALSE, covs = NULL)

Arguments

params
A "sampled_params" object returned from sample_params.

data
A data table of class "expanded_hesim_data" returned from create_transmod_data.

check_covs
Logical indicating whether to check that all covariates in data are contained in params.

covs
If check_covs is TRUE, then data_covs cannot be NULL and must specify all of the covariates in data that should be contained in params.

Value

A params_surv objects from the hesim package.

Details

The "dist" attribute from data is used to select a survival distribution from the mstate_nma element contained in params. The covariates for the selected survival distribution in mstate_nma that are also contained in data are extracted.

See also

Examples

# Joint distribution of parameters
params <- sample_params(n = 2)

# Treatment sequences
txseq1 <- txseq(first = "erlotinib",
                second = c("osimertinib", "PBDC"),
                second_plus = c("PBDC + bevacizumab", "PBDC + bevacizumab"))
txseq2 <- txseq(first = "gefitinib",
                second = c("osimertinib", "PBDC"),
                second_plus = c("PBDC + bevacizumab", "PBDC + bevacizumab"))
txseqs <- txseq_list(seq1 = txseq1, seq2 = txseq2)

# Patient population
pats <- create_patients(n = 2)

# Model structure
struct <- model_structure(txseqs, dist = "weibull")
tmat <- create_trans_mat(struct)

# Data and parameters for state transition model
transmod_data <- create_transmod_data(struct, tmat, pats, mutation_prob = .45)
transmod_params <- create_transmod_params(params, transmod_data)
print(transmod_params)

#> $coefs
#> $coefs$scale
#>      osi_s1p1_scale osi_s1d_scale osi_s2p2_scale pbdc_s2p2_scale osi_s2d_scale
#> [1,]      0.6608430     0.5349929       1.333809      -0.5191719   -0.02496359
#> [2,]      0.2106022     1.3980063       1.302304       0.9958771    1.63016274
#>      pbdc_s2d_scale osi_p2d_scale pbdc_p2d_scale d_erl_s1p1_scale
#> [1,]      0.2374274     0.2749992       1.097985       0.02223472
#> [2,]      0.5856988     1.2874032       2.099895      -0.08762537
#>      d_gef_s1p1_scale d_erl_s1d_scale d_gef_s1d_scale
#> [1,]       0.22171976      -0.1156124      -0.1925303
#> [2,]      -0.01380463       0.1464703      -0.1044621
#> 
#> $coefs$shape
#>      osi_s1p1_shape osi_s1d_shape osi_s2p2_shape pbdc_s2p2_shape osi_s2d_shape
#> [1,]      0.9878838     3.0125846     0.02157306        2.211227      1.039790
#> [2,]      1.5744670    -0.2903451     2.96957688        0.977201      1.373762
#>      pbdc_s2d_shape osi_p2d_shape pbdc_p2d_shape d_erl_s1p1_shape
#> [1,]       1.366512    -0.4644602       1.828704      -0.05280403
#> [2,]       1.197132    -2.4708456       2.409842      -0.03581339
#>      d_gef_s1p1_shape d_erl_s1d_shape d_gef_s1d_shape
#> [1,]      -0.05117418     -0.13636528     -0.06884081
#> [2,]       0.09202027     -0.05969474     -0.04915020
#> 
#> 
#> $dist
#> [1] "weibullNMA"
#> 
#> $n_samples
#> [1] 2
#> 
#> attr(,"class")
#> [1] "params_surv"