Create utility model

Create a model for health state utility given values of utility by health state, treatment, and time sampled from a probability distribution.

create_utilmod(n = 100, struct, patients, ae_probs,
  params_utility = iviNSCLC::params_utility, ae_duration = c("month",
  "progression"))

Arguments

n
The number of random observations of the parameters to draw.

struct
A model_structure object.

patients
A data table returned from create_patients.

ae_probs
An "ae_probs" object as returned by ae_probs.

params_utility
Parameter estimates for health state utilities and adverse event disutilities in the same format as params_utility.

ae_duration
Duration of time over with disutility from adverse events should accrue. If "month", then disutility accrues over the first month of treatment; if progression then disutility accrues until disease progression (i.e., over the entire duration of time spent in stable disease).

Value

An object of class "StateVals" from the hesim package.

Examples

# Treatment sequences
txseq1 <- txseq(first = "erlotinib",
                second = c("osimertinib", "PBDC"),
                second_plus = c("PBDC + bevacizumab", "PBDC + bevacizumab"))
txseq2 <- txseq(first = "gefitinib",
                second = c("osimertinib", "PBDC"),
                second_plus = c("PBDC + bevacizumab", "PBDC + bevacizumab"))
txseqs <- txseq_list(seq1 = txseq1, seq2 = txseq2)

# Patient population
pats <- create_patients(n = 2)

## Model structure
struct <- model_structure(txseqs, dist = "weibull")

## Utility model
n_samples <- 2
ae_probs <- ae_probs(n = n_samples, struct = struct)
utilmod <- create_utilmod(n = n_samples, struct = struct, patients = pats,
                          ae_probs = ae_probs)