Linear partial value function

Convert performance for criterion on original scale to a score on a common scale using a linear partial value function.

lpvf(x, x_min, x_max, score_min = 0, score_max = 100)

Arguments

x
Performance of the criterion on the original scale.

x_min
Minimum value of criterion on original scale. If higher performance is better, then the minimum value is the lowest possible value; otherwise, it is the highest.

x_max
Maximum value of criterion on original scale. If higher performance is better, then the maximum value is the highest possible value; otherwise, it is the lowest.

score_min
Minimum value of score on common scale.

score_max
Maximum value of score on common scale.

Value

The value of the score on the common scale.