Data to plot linear partial value function

Generate data used to plot a linear partial value function for a particular criteria given model outcomes.

lpvf_plot_data(x, optimal = c("low", "high"), length_out = 1000)

Arguments

x
Model outcomes for a particular criteria on the original scale.

optimal
If "low", then lower values of the outcome are better, and, if "high", then higher values of the outcome are better.

length_out
Number of points between minimum and maximum values of x.

Value

A data.table containing x and y coordinates for a line plot.

See also

Examples

outcome <- rnorm(10, mean = 100, sd = 11)
plot_data <- lpvf_plot_data(outcome, optimal = "high")
print(plot_data)

#>               x           y
#>    1:  91.44747   0.0000000
#>    2:  91.46943   0.1001001
#>    3:  91.49140   0.2002002
#>    4:  91.51336   0.3003003
#>    5:  91.53532   0.4004004
#>   ---                      
#>  996: 113.30213  99.5995996
#>  997: 113.32410  99.6996997
#>  998: 113.34606  99.7997998
#>  999: 113.36802  99.8998999
#> 1000: 113.38999 100.0000000