Multi criteria decision analysis

Conduct a multi criteria decision analysis (MCDA) and compute scores for competing treatment strategies using output from a probabilistic sensitivity analysis (PSA).

mcda(x, sample, strategy, criteria, criteria_min = NULL,
  criteria_max = NULL, optimal = NULL, weights, score_min = 0,
  score_max = 100)

Arguments

x
A data.frame or data.table of simulation output characterizing the probability distribution of model outcomes.

sample
Character name of column from x denoting a randomly sampled parameter set from the PSA.

strategy
Character name of column from x denoting treatment strategy.

criteria
A vector of character names of columns from x denoting the criteria to use in the MCDA.

criteria_min
A vector of minimum values for each criterion. If NULL, then the minimum value is computed automatically.

criteria_max
A vector of maximum values for each criterion. If NULL, then the maximum value is computed automatically.

optimal
A character vector denoting whether the optimal value of each criteria is "low" or "high". If an element is "low", then lower performance on that criterion is better, and, if an element is "high", then higher performance on that criterion is better. Must be specified if either criteria_min or criteria_min is NULL.

weights
Weights to apply to each criteria. Internally normalized to sum to 1.

score_min
Minimum of total value score. Default is 0.

score_max
Maximum of total value score. Default is 100.

See also

Examples

n_samples <- 5
strategies <- c("Strategy 1", "Strategy 2")
outcome1 <- c(rnorm(n_samples, mean = 10, sd = 5),
              rnorm(n_samples, mean = 8, sd = 4))
outcome2 <- c(rnorm(n_samples, mean = 1500, sd = 90),
              rnorm(n_samples, mean = 1000, sd = 100))
outcomes <- data.frame(sample = rep(1:n_samples, length(strategies)),
                       strategy_id = rep(strategies, each = n_samples),
                       criteria1 = outcome1,
                       criteria2 = outcome2)

# Performance matrix
performance_mat <- performance_matrix(outcomes,
                                      strategy = "strategy_id",
                                      criteria = c("criteria1", "criteria2"),
                                      rownames = c("Criteria 1", "Criteria 2"),
                                      colnames = strategies)
print(performance_mat)

#>            Strategy 1             Strategy 2           
#> Criteria 1 "12.59 (8.49, 15.31)"  "10.17 (6.71, 14.40)"
#> Criteria 2 "1,478 (1,407, 1,603)" "913 (833, 946)"     

# MCDA                        
weights <- c(.7, .3)
mcda <- mcda(outcomes, sample = "sample", strategy = "strategy_id",
             criteria = c("criteria1", "criteria2"),
             weights = weights,
             optimal = c("low", "high"))
names(mcda)

#> [1] "scores"          "weighted_scores" "total_value"     "prob_rank"      

# Scores on common scale
print(mcda$scores)

#>     sample strategy_id  criteria1 criteria2
#>  1:      1  Strategy 1  36.025162 100.00000
#>  2:      2  Strategy 1  82.735636  81.71461
#>  3:      3  Strategy 1  36.580656  80.68814
#>  4:      4  Strategy 1   0.000000  76.17513
#>  5:      5  Strategy 1   0.823855  73.10044
#>  6:      1  Strategy 2  85.514260  12.09776
#>  7:      2  Strategy 2  61.501686   0.00000
#>  8:      3  Strategy 2   7.145114  13.41999
#>  9:      4  Strategy 2 100.000000  15.08211
#> 10:      5  Strategy 2  40.422956  15.35761

# "Total value"
print(mcda$total_value)

#>     sample strategy_id     score rank
#>  1:      1  Strategy 1 55.217613    2
#>  2:      2  Strategy 1 82.429327    1
#>  3:      3  Strategy 1 49.812901    1
#>  4:      4  Strategy 1 22.852538    2
#>  5:      5  Strategy 1 22.506830    2
#>  6:      1  Strategy 2 63.489311    1
#>  7:      2  Strategy 2 43.051180    2
#>  8:      3  Strategy 2  9.027576    2
#>  9:      4  Strategy 2 74.524632    1
#> 10:      5  Strategy 2 32.903352    1

# "Total value" decomposed by criteria
print(mcda$weighted_scores)

#>     sample strategy_id  criteria weighted_score
#>  1:      1  Strategy 1 criteria1     25.2176132
#>  2:      2  Strategy 1 criteria1     57.9149450
#>  3:      3  Strategy 1 criteria1     25.6064593
#>  4:      4  Strategy 1 criteria1      0.0000000
#>  5:      5  Strategy 1 criteria1      0.5766985
#>  6:      1  Strategy 2 criteria1     59.8599822
#>  7:      2  Strategy 2 criteria1     43.0511799
#>  8:      3  Strategy 2 criteria1      5.0015796
#>  9:      4  Strategy 2 criteria1     70.0000000
#> 10:      5  Strategy 2 criteria1     28.2960694
#> 11:      1  Strategy 1 criteria2     30.0000000
#> 12:      2  Strategy 1 criteria2     24.5143819
#> 13:      3  Strategy 1 criteria2     24.2064413
#> 14:      4  Strategy 1 criteria2     22.8525378
#> 15:      5  Strategy 1 criteria2     21.9301317
#> 16:      1  Strategy 2 criteria2      3.6293286
#> 17:      2  Strategy 2 criteria2      0.0000000
#> 18:      3  Strategy 2 criteria2      4.0259965
#> 19:      4  Strategy 2 criteria2      4.5246325
#> 20:      5  Strategy 2 criteria2      4.6072827

# Probability of ranking
print(mcda$prob_rank)

#>    strategy_id rank N prob
#> 1:  Strategy 1    1 2  0.4
#> 2:  Strategy 1    2 3  0.6
#> 3:  Strategy 2    1 3  0.6
#> 4:  Strategy 2    2 2  0.4