Adverse event network meta-analysis parameters

The posterior distribution of the probability of an adverse event by available first and second line treatments. Based on separate models by adverse event with a binomial likelihood and a logit link.

params_ae_nma

Format

A list where each element is a matrix for a distinct adverse event. Each column is a distinct treatment and each row is a random draw from the posterior distribution. Each matrix contains a "imputed" attribute, which is vector of 0's and 1's with length equal to the number of columns, and a 1 denotes whether values of a column have been imputed because of insufficient clinical trial data.

When there is missing data for a TKI for a given adverse event, it is assumed to have the same adverse event probabilities as the TKI (among the TKI's with data) with the highest mean adverse event probability. When there is missing data for a combination therapy containing PBDC, it is assumed to have the same adverse event probabilities as PBDC. Finally, when there is not data for any PBDC based therapy, all PBDC based therapies are assumed to have the same adverse event probabilities as the TKI with the highest mean adverse event probability.

Examples

names(params_ae_nma)

#>  [1] "alt"          "ast"          "diarrhea"     "dry_skin"     "eye_problems"
#>  [6] "paronychia"   "pneumonitis"  "pruritus"     "rash"         "stomatitis"  

head(params_ae_nma[[1]])

#>          prob_erl     prob_gef     prob_afa     prob_dac    prob_osi
#> [1,] 0.0005113060 0.0003843076 3.317018e-05 2.775941e-05 0.001466678
#> [2,] 0.0006702925 0.0042614230 4.696339e-04 9.955454e-04 0.001373670
#> [3,] 0.0102076810 0.0039458165 9.924525e-04 7.983287e-04 0.007057089
#> [4,] 0.0115339318 0.0044082965 1.800498e-04 6.814098e-04 0.005149922
#> [5,] 0.0012178507 0.0018150712 1.013994e-04 6.576672e-04 0.006001942
#> [6,] 0.0005781604 0.0015465088 4.459036e-05 1.013073e-04 0.008285123
#>         prob_pbdc prob_pbdc_bev
#> [1,] 5.211729e-05  0.0005113060
#> [2,] 4.466606e-04  0.0006702925
#> [3,] 5.381788e-04  0.0102076810
#> [4,] 5.770781e-04  0.0115339318
#> [5,] 2.765439e-04  0.0012178507
#> [6,] 1.491408e-04  0.0005781604

attr(params_ae_nma[[1]], "imputed")

#> [1] 0 0 0 0 0 0 1