Adverse event cost parameters

Mean costs by adverse event.

params_costs_ae

Format

A data table with the following columns:

ae_name

Name of the adverse event.

ae_abb

Abbreviation for the adverse event.

mean

Mean costs.

lower

2.5% quantile for mean costs. Currently assumed to be +20% of mean costs given lack of data for estimating standard errors.

upper

97.5% quantile for mean costs. Currently assumed to be -20% of mean costs given lack of data for estimating standard errors.

se

Standard error of costs.

ref

Reference for mean costs.

Examples

print(params_costs_ae)

#>                             ae_name       ae_abb  mean     lower    upper
#>  1:                        Diarrhea     diarrhea 16510 14090.000 18791.00
#>  2:                        Dry skin     dry_skin   940   106.000  1844.00
#>  3:   Elevated alanine transaminase          alt 19122  7915.762 30328.24
#>  4: Elevated aspartate transaminase          ast 19122  7915.762 30328.24
#>  5:                    Eye problems eye_problems 31975 21923.000 49448.00
#>  6:                      Paronychia   paronychia  7788  6230.400  9345.60
#>  7:                     Pneumonitis  pneumonitis 21929 20697.000 23425.00
#>  8:                        Pruritus     pruritus 26172  3663.000 62005.00
#>  9:                            Rash         rash 15709 11315.000 21373.00
#> 10:                      Stomatitis   stomatitis 18151 15111.000 22542.00
#>             se                ref
#>  1:  1199.2347 wong2018assessment
#>  2:   443.3673 wong2018assessment
#>  3:  5717.4686  bilir2016economic
#>  4:  5717.4686  bilir2016economic
#>  5:  7021.6837 wong2018assessment
#>  6:   794.6939            DRG 602
#>  7:   695.9184 wong2018assessment
#>  8: 14883.1633 wong2018assessment
#>  9:  2565.8163 wong2018assessment
#> 10:  1895.6633 wong2018assessment
#>                                                         notes
#>  1: Grade 3/4 from trials, severe AE (i.e. inpatient) assumed
#>  2:              Used the cost estimate for dermatitis all AE
#>  3: Grade 3/4 from trials, severe AE (i.e. inpatient) assumed
#>  4: Grade 3/4 from trials, severe AE (i.e. inpatient) assumed
#>  5: Grade 3/4 from trials, severe AE (i.e. inpatient) assumed
#>  6:                                                      <NA>
#>  7: Grade 3/4 from trials, severe AE (i.e. inpatient) assumed
#>  8: Grade 3/4 from trials, severe AE (i.e. inpatient) assumed
#>  9: Grade 3/4 from trials, severe AE (i.e. inpatient) assumed
#> 10: Grade 3/4 from trials, severe AE (i.e. inpatient) assumed