Productivity cost parameters

Parameters used to estimate productivity costs including costs from temporary disability, permanent disability, and premature mortality.

params_costs_prod

Format

A list with the following elements:

wages

A data.table with columns:

gender

Male or female gender.

employment_status

Employed full-time, employed part-time, or unemployed.

prop

Proportion of population.

weekly_wage

Weekly wage.

source_prop

Source for prop.

source_wage

Source for wage.

Mean wages are computed as a weighted average of wages for full-time, part-time, and unemployed workers.

temporary_disability

A vector with three elements:

missed_days_est

Estimate for number of days worked following diagnosis due to sick leave and leave of absence.

missed_days_lower

Lower bound for missed_days_est.

missed_days_upper

Upper bound for missed_days_upper.

permanent_disability

A vector with three elements:

hours_reduction_est

Estimate of reduced working hours following initial abscence.

hours_reduction_lower

Lower bound for hours_reduction_est.

hours_reduction_upper

Upper bound for hours_reduction_est.

Examples

print(params_costs_prod)

#> $wages
#>    gender employment_status       prop weekly_wage
#> 1: female              full 0.72360125         796
#> 2: female              part 0.23331205         326
#> 3: female        unemployed 0.04308671           0
#> 4:   male              full 0.84057784         973
#> 5:   male              part 0.11546186         321
#> 6:   male        unemployed 0.04396030           0
#>                             source_prop
#> 1: https://www.bls.gov/cps/cpsaat08.htm
#> 2: https://www.bls.gov/cps/cpsaat08.htm
#> 3: https://www.bls.gov/cps/cpsaat08.htm
#> 4: https://www.bls.gov/cps/cpsaat08.htm
#> 5: https://www.bls.gov/cps/cpsaat08.htm
#> 6: https://www.bls.gov/cps/cpsaat08.htm
#>                                        source_wage
#> 1: https://www.bls.gov/news.release/pdf/wkyeng.pdf
#> 2: https://www.bls.gov/news.release/pdf/wkyeng.pdf
#> 3:                                            <NA>
#> 4: https://www.bls.gov/news.release/pdf/wkyeng.pdf
#> 5: https://www.bls.gov/news.release/pdf/wkyeng.pdf
#> 6:                                            <NA>
#> 
#> $temporary_disability
#>   missed_days_est missed_days_lower missed_days_upper 
#>             151.0             120.8             181.2 
#> 
#> $permanent_disability
#>   hours_reduction_est hours_reduction_lower hours_reduction_upper 
#>                     4                     3                     5 
#>