Treatment cost parameters

Information on dosage, strength, pricing, and administration costs needed to estimate costs of treatments sequences.

params_costs_tx

Format

A list with the following elements:

dosage

A data.table containing information on dosage for each agent used for treatments that can be included in a treatment sequence. The variables strength1, strength2, quantity1, quantity2, and units_per_day can be modified to change annualized costs.

agent_name

Name of the agent used in the treatment; combination therapies will have more than one agents.

dosage

Amount, number. and frequency of doses over a specified time period.

dose

Specified amount of medication taken at one time. By default, doses based on body surface area (BSA) and weight are based on mean BSA and weight, respectively.

unit

The unit used to compute costs.

strength1

The amount of the drug in a single dosage form (e.g., tablet or vial).

strength2

If a unit consists of multiple dosage forms, then the amount of the drug in the dosage form not included in strength1.

quantity1

Quantity of dosage form specified in strength1 in a unit.

quantity2

Quantity of dosage form specified in strength2 in a unit.

units_per_day

Number of units of the treatment (based on dosage) used per day.

duration_days

Number of days to use the treatment. If used until progression, then must equal Inf.

source

Source for dosage information.

acquisition_costs

A data.table of acquisition costs by agent and strength. Default costs are based on wholesale acquisition costs (WACs). The variable acquisition_cost can be modified to change annualized acquisition costs.

agent_name

Name of the agent

strength

The amount of the agent in a given unit of the dosage form.

acquisition_cost

The acquisition cost of an agent for a given strength.

source

Source used for acquisition cost.

discounts

agent_name

Name of the agent.

discount_lower

Lower bound for discount and rebates as a proportion of acquisition_cost in the acquisition_costs table. Assumed to follow a uniform distribution in the probabilistic sensitivity analysis.

discount_upper

Upper bound for discount and rebates as a proportion of acquisition_cost in the acquisition_costs table. Assumed to follow a uniform distribution in the probabilistic sensitivity analysis.

administration_costs

A data.table of administration costs. The variable costs can be modified to change annualized administration costs.

agent_name

Name of the agent.

administration_cost

The costs of administering an agent.

source

Source used for administration cost.

lookup

A lookup data.table used to match treatments from the treatments table to agents comprising a particular treatment.

tx_name

Name of treatment.

agent_name1

Name of first agent used for treatment.

agent_name2

Name of second agent used for treatment.

agent_name3

Name of third agent used for treatment.

agent_name4

Name of fourth agent used for treatment.

Examples

print(params_costs_tx)

#> $dosage
#>        agent_name                                               dosage    dose
#>  1:     erlotinib                            150 mg orally, once daily  150.00
#>  2:     gefitinib                                 250 mg orally, daily  250.00
#>  3:      afatinib                             40 mg orally, once daily   40.00
#>  4:   dacomitinib                             45 mg orally, once daily   45.00
#>  5:   osimertinib                             80 mg orally, once daily   80.00
#>  6:     cisplatin                   75mg/m2, 1x/cycle, 6 21-day cycles  142.50
#>  7:    pemetrexed                    500 mg, 1x/cycle, 6 21-day cycles  500.00
#>  8:   bevacizumab 15mg/kg IV every 3 weeks with carboplatin/paclitaxel 1111.95
#>  9:     nivolumab         240 mg every 2 weeks or 480 mg every 4 weeks  240.00
#> 10: pembrolizumab                                 200 mg every 3 weeks  200.00
#> 11:  atezolizumab  1200 mg as an IV infusion over 60 min every 3 weeks 1200.00
#>                                     unit        strength1      strength2
#>  1:                         150mg tablet     150mg tablet           <NA>
#>  2:                         250mg tablet     250mg tablet           <NA>
#>  3:                          40mg tablet      40mg tablet           <NA>
#>  4:                          45mg tablet      45mg tablet           <NA>
#>  5:                          80mg tablet      80mg tablet           <NA>
#>  6: 1 100mg/100ml vial, 1 50mg/50ml vial 100mg/100ml vial 50mg/50ml vial
#>  7:                           500mg vial       500mg vial           <NA>
#>  8:                   3 400mg/16ml vials  400mg/16ml vial           <NA>
#>  9:             2 100mg/10ml, 1 40mg/4ml  100mg/10ml vial  40mg/4ml vial
#> 10:                     2 100mg/4ml vial   100mg/4ml vial           <NA>
#> 11:                     1200mg/20ml vial 1200mg/20ml vial           <NA>
#>     quantity1 quantity2 units_per_day duration_days
#>  1:         1         0    1.00000000           Inf
#>  2:         1         0    1.00000000           Inf
#>  3:         1         0    1.00000000           Inf
#>  4:         1         0    1.00000000           Inf
#>  5:         1         0    1.00000000           Inf
#>  6:         1         1    0.04761905           126
#>  7:         1         0    0.04761905           126
#>  8:         3         0    0.04761905           Inf
#>  9:         2         1    0.07142857           730
#> 10:         2         0    0.04761905           730
#> 11:         1         0    0.04761905           730
#>                                                                                                                                  source
#>  1:                                                                                                                           FDA label
#>  2:                                                                                                                           FDA label
#>  3:                                                                                                                           FDA label
#>  4:                                                                                                                           FDA label
#>  5:                                                                                                                           FDA label
#>  6: ICER Report Treatment Options for Advanced Non-Small Cell Lung Cancer: Effectiveness, Value and Value Based Price Benchmarks (2016)
#>  7:                                                                                                                                <NA>
#>  8:                                                                                                                           FDA label
#>  9:                                                                                                                           FDA label
#> 10:                                                                                                                           FDA label
#> 11:                                                                                                                           FDA label
#> 
#> $acquisition_costs
#>        agent_name         strength acquisition_cost
#>  1:     erlotinib     150mg tablet         281.7133
#>  2:     gefitinib     250mg tablet         257.0590
#>  3:      afatinib      40mg tablet         235.0500
#>  4:   dacomitinib      45mg tablet         407.3900
#>  5:   osimertinib      80mg tablet         487.2157
#>  6:     cisplatin   50mg/50ml vial          14.5100
#>  7:     cisplatin 100mg/100ml vial          35.0000
#>  8:     cisplatin 200mg/200ml vial         100.0000
#>  9:    pemetrexed       100mg vial         676.5200
#> 10:    pemetrexed       500mg vial        3382.6000
#> 11:   carboplatin         50mg/5ml           5.9900
#> 12:   carboplatin       150mg/15ml          14.2000
#> 13:   carboplatin       450mg/45ml          32.9900
#> 14:   carboplatin       600mg/60ml          43.9900
#> 15:   gemcitabine        200mg/2mL           3.7000
#> 16:   gemcitabine          1g/10mL           3.6900
#> 17:   gemcitabine          2g/20mL           4.1000
#> 18:     docetaxel       160mg/16ml          34.3800
#> 19:     docetaxel         80mg/8ml          34.3800
#> 20:     docetaxel         20mg/2ml          34.3800
#> 21:    paclitaxel       300mg/50ml          55.0000
#> 22:    paclitaxel     100mg/16.7ml          24.5000
#> 23:    paclitaxel         30mg/5ml          10.5000
#> 24:   bevacizumab   100mg/4ml vial         199.2350
#> 25:   bevacizumab  400mg/16ml vial         199.2350
#> 26:     nivolumab  100mg/10ml vial         262.2080
#> 27:     nivolumab    40mg/4ml vial         262.2075
#> 28: pembrolizumab   100mg/4ml vial        1162.4100
#> 29:  atezolizumab 1200mg/20ml vial         444.0270
#>        agent_name         strength acquisition_cost
#>                                      source
#>  1:       ProspectoRx; WAC price 12/15/2017
#>  2:         ProspectoRx; WAC price 1/1/2018
#>  3:         ProspectoRx; WAC price 9/1/2017
#>  4:                     Public announcement
#>  5:         ProspectoRx; WAC price 1/1/2018
#>  6:         ProspectoRx; WAC price 3/8/2018
#>  7:        ProspectoRx; WAC price 9/12/2016
#>  8:       ProspectoRx; WAC price 10/17/2016
#>  9:        ProspectoRx; WAC price 3/30/2018
#> 10:        ProspectoRx; WAC price 3/30/2018
#> 11: AnalySource 201875 (NDC: 47781-0603-20)
#> 12: AnalySource 201875 (NDC: 50742-0446-15)
#> 13: AnalySource 201875 (NDC: 47781-0605-94)
#> 14: AnalySource 201875 (NDC: 47781-0606-94)
#> 15:         ProspectoRx; WAC price 5/1/2018
#> 16:         ProspectoRx; WAC price 5/1/2018
#> 17:         ProspectoRx; WAC price 5/1/2018
#> 18:         ProspectoRx; WAC price 9/5/2018
#> 19:        ProspectoRx; WAC price 9/28/2018
#> 20:        ProspectoRx; WAC price 9/28/2018
#> 21:        ProspectoRx; WAC price 1/23/2018
#> 22:        ProspectoRx; WAC price 1/23/2018
#> 23:        ProspectoRx; WAC price 1/23/2018
#> 24:         ProspectoRx; WAC price 7/1/2018
#> 25:         ProspectoRx; WAC price 7/1/2018
#> 26:         ProspectoRx; WAC price 4/1/2018
#> 27:         ProspectoRx; WAC price 4/1/2018
#> 28:        ProspectoRx; WAC price 3/22/2018
#> 29:         ProspectoRx; WAC price 7/1/2018
#>                                      source
#> 
#> $discounts
#>        agent_name discount_lower discount_upper
#>  1:     erlotinib            0.2            0.3
#>  2:     gefitinib            0.2            0.3
#>  3:      afatinib            0.2            0.3
#>  4:   dacomitinib            0.2            0.3
#>  5:   osimertinib            0.2            0.3
#>  6:     cisplatin            0.2            0.3
#>  7:    pemetrexed            0.2            0.3
#>  8:   bevacizumab            0.2            0.3
#>  9:     nivolumab            0.2            0.3
#> 10: pembrolizumab            0.2            0.3
#> 11:  atezolizumab            0.2            0.3
#> 
#> $administration_costs
#>        agent_name administration_cost                  source
#>  1:     erlotinib                0.00                     N/A
#>  2:     gefitinib                0.00                     N/A
#>  3:      afatinib                0.00                     N/A
#>  4:   dacomitinib                0.00                     N/A
#>  5:   osimertinib                0.00                     N/A
#>  6:     cisplatin               91.72        CPT 96417, 96415
#>  7:    pemetrexed              136.15               CPT 96413
#>  8:   bevacizumab               91.72 CPT 96413, 96415, 96417
#>  9:     nivolumab              136.15               CPT 96413
#> 10: pembrolizumab              136.15               CPT 96413
#> 11:  atezolizumab              136.15               CPT 96413
#> 
#> $lookup
#>                                tx_name   agent_name1 agent_name2   agent_name3
#>  1:                          erlotinib     erlotinib        <NA>          <NA>
#>  2:                          gefitinib     gefitinib        <NA>          <NA>
#>  3:                           afatinib      afatinib        <NA>          <NA>
#>  4:                        dacomitinib   dacomitinib        <NA>          <NA>
#>  5:                        osimertinib   osimertinib        <NA>          <NA>
#>  6:                               PBDC     cisplatin  pemetrexed          <NA>
#>  7:                        bevacizumab   bevacizumab        <NA>          <NA>
#>  8:                          nivolumab     nivolumab        <NA>          <NA>
#>  9:                      pembrolizumab pembrolizumab        <NA>          <NA>
#> 10:                       atezolizumab  atezolizumab        <NA>          <NA>
#> 11:                 PBDC + bevacizumab     cisplatin  pemetrexed   bevacizumab
#> 12:                   PBDC + nivolumab     cisplatin  pemetrexed     nivolumab
#> 13:               PBDC + pembrolizumab     cisplatin  pemetrexed pembrolizumab
#> 14:                PBDC + atezolizumab     cisplatin  pemetrexed  atezolizumab
#> 15:     PBDC + bevacizumab + nivolumab     cisplatin  pemetrexed   bevacizumab
#> 16: PBDC + bevacizumab + pembrolizumab     cisplatin  pemetrexed   bevacizumab
#> 17:  PBDC + bevacizumab + atezolizumab     cisplatin  pemetrexed   bevacizumab
#>       agent_name4
#>  1:          <NA>
#>  2:          <NA>
#>  3:          <NA>
#>  4:          <NA>
#>  5:          <NA>
#>  6:          <NA>
#>  7:          <NA>
#>  8:          <NA>
#>  9:          <NA>
#> 10:          <NA>
#> 11:          <NA>
#> 12:          <NA>
#> 13:          <NA>
#> 14:          <NA>
#> 15:     nivolumab
#> 16: pembrolizumab
#> 17:  atezolizumab
#>