Utility parameters

Utility estimates by health state and disutilities by adverse event.

params_utility

Format

A list containing the following elements:

state_utility Utility estimates by health state.

as_disutility Disutility by health state.

State utility

The state_utility element is a data table with the following columns:

state_name

Name of the health state.

mean

Mean utility.

se

Standard error of utility.

ref

BibTeX reference for estimate.

Adverse event disutilities

The as_disutility element is a data table with the following columns:

ae_name

Name of the adverse event.

ae_abb

Abbreviation for the adverse event.

mean

Mean disutility.

se

Standard error of disutility.

ref

BibTeX reference for estimate.

Examples

print(params_utility)

#> $state_utility
#>    state_name   mean         se              ref
#> 1:         S1 0.7540 0.00000000 nafees2017health
#> 2:      P1/S2 0.6532 0.02223000 nafees2008health
#> 3:         P2 0.4734 0.03105848 nafees2008health
#> 
#> $ae_disutility
#>                             ae_name       ae_abb  mean    se              ref
#>  1:   Elevated alanine transaminase          alt 0.000 0.000             <NA>
#>  2: Elevated aspartate transaminase          ast 0.000 0.000             <NA>
#>  3:                        Diarrhea     diarrhea 0.222 0.073 nafees2017health
#>  4:                        Dry skin     dry_skin 0.151 0.050 nafees2017health
#>  5:                    Eye problems eye_problems 0.138 0.046 nafees2017health
#>  6:                      Paronychia   paronychia 0.151 0.050 nafees2017health
#>  7:                     Pneumonitis  pneumonitis 0.165 0.054  doyle2008health
#>  8:                        Pruritus     pruritus 0.151 0.050 nafees2017health
#>  9:                            Rash         rash 0.151 0.050 nafees2017health
#> 10:                      Stomatitis   stomatitis 0.151 0.050 nafees2017health
#>                                                                   notes
#>  1:                                                          Assumption
#>  2:                                                          Assumption
#>  3:                                                                <NA>
#>  4:          Assumed equal to rash disutility; se assumed 1/3rd of mean
#>  5:                Assumed equal to hair loss; se assumed 1/3rd of mean
#>  6:          Assumed equal to rash disutility; se assumed 1/3rd of mean
#>  7: Assumed equal to cough, dyspnea, and pain; se assumed 1/3rd of mean
#>  8:          Assumed equal to rash disutility; se assumed 1/3rd of mean
#>  9:                                                                <NA>
#> 10:          Assumed equal to rash disutility; se assumed 1/3rd of mean
#>