Sample parameters for PSA

Sample parameters from their joint probability distribution for a probabilistic sensitivity analysis (PSA).

sample_params(n, params_mstate_nma = iviNSCLC::params_mstate_nma,
  params_utility = iviNSCLC::params_utility)

Arguments

n
The number of random observations of the parameters to draw.

params_mstate_nma
A list of params_surv objects, where each element in the list denotes a survival distribution. Should have the same variable names as params_mstate_nma.

params_utility
Parameter estimates for health state utilities and adverse event disutilities in the same format as params_utility.

mstate_nma

A params_surv object where the number of rows in each coef element is equal to n.

utility

A 3-dimensional array of matrices where each matrix represents the utility values for a different treatment. The rows of each matrix are random draws from a beta distribution and columns are health states.

Value

An object of class "sampled_params", which is a list with the following components.

Examples

params <- sample_params(2)
names(params)

#> [1] "mstate_nma" "utility"   

# Multi-state NMA
params$mstate_nma$weibull$coefs$shape

#>      osi_s1p1_scale d_erl_s1p1_scale d_gef_s1p1_scale d_afa_s1p1_scale
#> [1,]              0                0                0                0
#> [2,]              0                0                0                0
#>      d_dac_s1p1_scale osi_s1d_scale d_erl_s1d_scale d_gef_s1d_scale
#> [1,]                0             0               0               0
#> [2,]                0             0               0               0
#>      d_afa_s1d_scale d_dac_s1d_scale osi_p1d_scale d_erl_p1d_scale
#> [1,]               0               0             0               0
#> [2,]               0               0             0               0
#>      d_gef_p1d_scale d_afa_p1d_scale d_dac_p1d_scale osi_s2p2_scale
#> [1,]               0               0               0              0
#> [2,]               0               0               0              0
#>      pbdc_s2p2_scale d_bev_s2p2_scale d_pbdc_bev_s2p2_scale d_erl_s2p2_scale
#> [1,]               0                0                     0                0
#> [2,]               0                0                     0                0
#>      d_gef_s2p2_scale d_afa_s2p2_scale d_dac_s2p2_scale osi_s2d_scale
#> [1,]                0                0                0             0
#> [2,]                0                0                0             0
#>      pbdc_s2d_scale d_bev_s2d_scale d_pbdc_bev_s2d_scale d_erl_s2d_scale
#> [1,]              0               0                    0               0
#> [2,]              0               0                    0               0
#>      d_gef_s2d_scale d_afa_s2d_scale d_dac_s2d_scale osi_p2d_scale
#> [1,]               0               0               0             0
#> [2,]               0               0               0             0
#>      pbdc_p2d_scale d_bev_p2d_scale d_pbdc_bev_p2d_scale d_erl_p2d_scale
#> [1,]              0               0                    0               0
#> [2,]              0               0                    0               0
#>      d_gef_p2d_scale d_afa_p2d_scale d_dac_p2d_scale
#> [1,]               0               0               0
#> [2,]               0               0               0

params$mstate_nma$weibull$dist

#> [1] "weibull"