A list of treatment sequences

Create a list of objects of class "txseq".

txseq_list(..., start_line = c("first", "second"),
  mutation = c("unknown", "positive", "negative"))

Arguments

...
Objects to form a list.

start_line
The starting line of treatmnet that is being modeled. When modeling second line treatment, the first line must be specified in order to characterize a treatment history; however, note that disease transition model cannot currently be simulated for models starting at second line because the evidence base is too limited.

mutation
T790M mutation status. If the model starts at first line, then this must be "unknown"; otherwise, if the model starts at second line, then the mutation status must be specified as either "positive" or "negative", since it will be known.

Value

An object of class "txseq_list", which is a list of objects of class. start_line and mutation are stored as attributes.

See also

Examples

txseq1 <- txseq(first = "erlotinib",
               second = c("osimertinib", "PBDC"),
               second_plus = c("PBDC + bevacizumab", "PBDC + bevacizumab"))
txseq2 <- txseq(first = "gefitinib",
               second = c("osimertinib", "PBDC"),
               second_plus = c("PBDC + bevacizumab", "PBDC + bevacizumab"))
txseqs <- txseq_list(seq1 = txseq1, seq2 = txseq2)
class(txseqs)

#> [1] "txseq_list"

print(txseqs$seq1)

#> $first
#> [1] "erlotinib"
#> 
#> $second
#>           pos           neg 
#> "osimertinib"        "PBDC" 
#> 
#> $second_plus
#>                  pos                  neg 
#> "PBDC + bevacizumab" "PBDC + bevacizumab" 
#> 
#> attr(,"class")
#> [1] "txseq"

attributes(txseqs)

#> $names
#> [1] "seq1" "seq2"
#> 
#> $class
#> [1] "txseq_list"
#> 
#> $start_line
#> [1] "first"
#> 
#> $mutation
#> [1] "unknown"
#>