Parameters for transition model

Extract parameters from a multi-state NMA for use with the data table returned by create_transmod_data, which are used to simulate health state transitions with a continuous time state transition model (CTSTM).

create_transmod_params(n = 100, data,
  params_mstate_nma = iviNSCLC::params_mstate_nma, check_covs = FALSE,
  covs = NULL)

Arguments

n
The number of random observations of the parameters to draw.

data
A data table of class "expanded_hesim_data" returned from create_transmod_data.

params_mstate_nma
A list of params_surv objects, where each element in the list denotes a survival distribution. Should have the same variable names as params_mstate_nma.

check_covs
Logical indicating whether to check that all covariates in data are contained in params.

covs
If check_covs is TRUE, then data_covs cannot be NULL and must specify all of the covariates in data that should be contained in params.

Value

A params_surv objects from the hesim package.

Details

The "dist" attribute from data is used to select a survival distribution from the mstate_nma element contained in params. The covariates for the selected survival distribution in mstate_nma that are also contained in data are extracted.

See also