Adverse event cost parameters

Mean costs by adverse event.

params_costs_ae

Format

A data table with the following columns:

ae_name

Name of the adverse event.

ae_abb

Abbreviation for the adverse event.

mean

Mean costs.

lower

2.5% quantile for mean costs. Currently assumed to be +20% of mean costs given lack of data for estimating standard errors.

upper

97.5% quantile for mean costs. Currently assumed to be -20% of mean costs given lack of data for estimating standard errors.

se

Standard error of costs.

ref

Reference for mean costs.

Examples

print(params_costs_ae)

#>                             ae_name       ae_abb      mean      lower     upper
#>  1:                        Diarrhea     diarrhea 18011.018 15371.0020 20499.397
#>  2:                        Dry skin     dry_skin  1025.461   115.6371  2011.649
#>  3:   Elevated alanine transaminase          alt 21411.498  8863.5245 33959.472
#>  4: Elevated aspartate transaminase          ast 21411.498  8863.5245 33959.472
#>  5:                    Eye problems eye_problems 34882.029 23916.1446 53943.599
#>  6:                      Paronychia   paronychia  8249.013  6599.2100  9898.815
#>  7:                     Pneumonitis  pneumonitis 23922.690 22578.6820 25554.700
#>  8:                        Pruritus     pruritus 28551.445  3996.0242 67642.227
#>  9:                            Rash         rash 17137.195 12343.7110 23316.141
#> 10:                      Stomatitis   stomatitis 19801.211 16484.8270 24591.421
#>             se year                ref
#>  1:  1308.2639 2015 wong2018assessment
#>  2:   483.6764 2015 wong2018assessment
#>  3:  6402.0275 2014  bilir2016economic
#>  4:  6402.0275 2014  bilir2016economic
#>  5:  7660.0649 2015 wong2018assessment
#>  6:   841.7360 2016            DRG 602
#>  7:   759.1883 2015 wong2018assessment
#>  8: 16236.2763 2015 wong2018assessment
#>  9:  2799.0893 2015 wong2018assessment
#> 10:  2068.0088 2015 wong2018assessment
#>                                                         notes
#>  1: Grade 3/4 from trials, severe AE (i.e. inpatient) assumed
#>  2:              Used the cost estimate for dermatitis all AE
#>  3: Grade 3/4 from trials, severe AE (i.e. inpatient) assumed
#>  4: Grade 3/4 from trials, severe AE (i.e. inpatient) assumed
#>  5: Grade 3/4 from trials, severe AE (i.e. inpatient) assumed
#>  6:                                                      <NA>
#>  7: Grade 3/4 from trials, severe AE (i.e. inpatient) assumed
#>  8: Grade 3/4 from trials, severe AE (i.e. inpatient) assumed
#>  9: Grade 3/4 from trials, severe AE (i.e. inpatient) assumed
#> 10: Grade 3/4 from trials, severe AE (i.e. inpatient) assumed