Sample parameters for PSA

Sample parameters from their joint probability distribution for a probabilistic sensitivity analysis (PSA).

sample_params(n, params_mstate_nma = iviNSCLC::params_mstate_nma)

Arguments

n
The number of random observations of the parameters to draw.

params_mstate_nma
A list of params_surv objects, where each element in the list denotes a survival distribution. Should have the same variable names as params_mstate_nma.

Value

An object of class "sampled_params", which is a list with the following components.

mstate_nma

A params_surv object where the number of rows in each coef element is equal to n.

Examples

params <- sample_params(2)
names(params)

#> [1] "mstate_nma"

# Multi-state NMA
params$mstate_nma$weibull$coefs$shape

#>      osi_s1p1_shape osi_s1d_shape osi_p1d_shape osi_s2p2_shape pbdc_s2p2_shape
#> [1,]      0.2144393     0.8312884      1.378093     0.08360727      -0.7072081
#> [2,]      1.3830202     0.2256768     -2.038570    -1.18088236       1.1533245
#>      osi_s2d_shape pbdc_s2d_shape osi_p2d_shape pbdc_p2d_shape d_erl_s1p1_shape
#> [1,]     0.4761633       1.207093     -1.092998       1.275491       0.06605832
#> [2,]    -0.6254911       1.669549      1.763126       2.514795      -0.10613347
#>      d_gef_s1p1_shape d_afa_s1p1_shape d_dac_s1p1_shape d_erl_s1d_shape
#> [1,]      -0.07301225      -0.23635251       0.11426602      0.02534214
#> [2,]      -0.02748377      -0.05647148       0.06237375      0.25451736
#>      d_gef_s1d_shape d_afa_s1d_shape d_dac_s1d_shape d_erl_p1d_shape
#> [1,]     -0.05378574     -0.04179779        0.041887      -0.1066887
#> [2,]     -0.19719023     -0.05537505       -0.214220      -0.2297192
#>      d_gef_p1d_shape d_afa_p1d_shape d_dac_p1d_shape d_bev_s2p2_shape
#> [1,]      -0.1340769     -0.04057471     0.113551764       0.07741805
#> [2,]       0.3401152      0.15783271    -0.001641351       0.11573567
#>      d_pbdc_bev_s2p2_shape d_erl_s2p2_shape d_gef_s2p2_shape d_afa_s2p2_shape
#> [1,]             0.1265985       0.09302601        0.1432585        0.3017123
#> [2,]            -0.1774899       0.09931315        0.1632356       -0.2248138
#>      d_dac_s2p2_shape d_bev_s2d_shape d_pbdc_bev_s2d_shape d_erl_s2d_shape
#> [1,]      -0.19229882     -0.19171508           -0.2458048     -0.06971819
#> [2,]      -0.03614562      0.01928407            0.1163636     -0.13729794
#>      d_gef_s2d_shape d_afa_s2d_shape d_dac_s2d_shape d_bev_p2d_shape
#> [1,]     -0.05061834      -0.1425638      -0.1002112     -0.38684132
#> [2,]      0.03711062       0.2181432       0.2315258     -0.06762784
#>      d_pbdc_bev_p2d_shape d_erl_p2d_shape d_gef_p2d_shape d_afa_p2d_shape
#> [1,]          -0.33339637     -0.22122720     -0.03348921       0.2772291
#> [2,]           0.09517852     -0.08635971     -0.06537331       0.0975116
#>      d_dac_p2d_shape
#> [1,]      -0.3314919
#> [2,]       0.2705962

params$mstate_nma$weibull$dist

#> [1] "weibullNMA"