Sample parameters for PSA

Sample parameters from their joint probability distribution for a probabilistic sensitivity analysis (PSA).

sample_params(n, params_mstate_nma = iviNSCLC::params_mstate_nma)

Arguments

n
The number of random observations of the parameters to draw.

params_mstate_nma
A list of params_surv objects, where each element in the list denotes a survival distribution. Should have the same variable names as params_mstate_nma.

Value

An object of class "sampled_params", which is a list with the following components.

mstate_nma

A params_surv object where the number of rows in each coef element is equal to n.

Examples

params <- sample_params(2)
names(params)

#> [1] "mstate_nma"

# Multi-state NMA
params$mstate_nma$weibull$coefs$shape

#>      osi_s1p1_shape osi_s1d_shape osi_p1d_shape osi_s2p2_shape pbdc_s2p2_shape
#> [1,]      0.9878838     3.0125846     0.9438242     0.02157306        2.211227
#> [2,]      1.5744670    -0.2903451     2.3826415     2.96957688        0.977201
#>      osi_s2d_shape pbdc_s2d_shape osi_p2d_shape pbdc_p2d_shape d_erl_s1p1_shape
#> [1,]      1.039790       1.366512    -0.4644602       1.828704      -0.05280403
#> [2,]      1.373762       1.197132    -2.4708456       2.409842      -0.03581339
#>      d_gef_s1p1_shape d_afa_s1p1_shape d_dac_s1p1_shape d_erl_s1d_shape
#> [1,]      -0.05117418      -0.07887960      -0.06085395     -0.13636528
#> [2,]       0.09202027       0.09006009      -0.12086765     -0.05969474
#>      d_gef_s1d_shape d_afa_s1d_shape d_dac_s1d_shape d_erl_p1d_shape
#> [1,]     -0.06884081      -0.1269187     -0.09014325      0.04128342
#> [2,]     -0.04915020       0.1712486      0.16226955      0.16094859
#>      d_gef_p1d_shape d_afa_p1d_shape d_dac_p1d_shape d_bev_s2p2_shape
#> [1,]      -0.2434193     -0.08101871    -0.003653794       0.05079327
#> [2,]       0.1548038     -0.12905734    -0.058895599      -0.11811305
#>      d_pbdc_bev_s2p2_shape d_erl_s2p2_shape d_gef_s2p2_shape d_afa_s2p2_shape
#> [1,]            0.17294046      -0.16734310       0.06969866      -0.04098802
#> [2,]           -0.05339682       0.06241384       0.01873162       0.22311779
#>      d_dac_s2p2_shape d_bev_s2d_shape d_pbdc_bev_s2d_shape d_erl_s2d_shape
#> [1,]       -0.2105596      -0.1677768            0.1903635      -0.3365152
#> [2,]       -0.2292437       0.1578390           -0.0797747      -0.3021977
#>      d_gef_s2d_shape d_afa_s2d_shape d_dac_s2d_shape d_bev_p2d_shape
#> [1,]     0.002935077     -0.02492581      -0.3242136       0.2352541
#> [2,]    -0.038658105     -0.19264739      -0.2084033      -0.2361561
#>      d_pbdc_bev_p2d_shape d_erl_p2d_shape d_gef_p2d_shape d_afa_p2d_shape
#> [1,]          -0.05182098      0.01224905       0.1501016      -0.1452805
#> [2,]          -0.17623775     -0.10201078       0.1327201      -0.2216938
#>      d_dac_p2d_shape
#> [1,]      0.02387482
#> [2,]     -0.23560772

params$mstate_nma$weibull$dist

#> [1] "weibullNMA"